Update with new script structure
