+++ example-app-angular-cli/package.json
@@ -5,9 +5,10 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "playground:fresh": "node ../scripts/build-examples.js && npm run playground",
-    "playground": "node ./node_modules/angular-playground/dist/bin/cli.js",
-    "playground2": "node ./node_modules/angular-playground/dist/bin/cli.js angular-playground2.json",
+    "playground:copy": "node ../scripts/copy.js",
+    "playground:run": "node ./node_modules/angular-playground/dist/bin/cli.js",
+    "playground": "npm run playground:copy && npm run playground:run",
+    "playground2:run": "node ./node_modules/angular-playground/dist/bin/cli.js angular-playground2.json",
     "cli": "ts-node --project ../src/cli ../src/cli/cli.ts",
     "cli:build": "ts-node --project ../src/cli ../src/cli/cli.ts -no-watch -no-serve",
     "start": "ng serve -no-progress",
